flaskr/api/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
marshmallow==3.19.0
apispec==6.3.0
apispec-webframeworks==0.5.2
flasgger==0.9.5
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary
marshmallow-sqlalchemy